+++ noodles-sam/src/record.rs
@@ -541,14 +541,12 @@ impl AlignmentRecord for Record {
     /// # Examples
     ///
     /// ```
-    /// # use std::io;
     /// use noodles_sam::{self as sam, AlignmentRecord};
     /// let record = sam::Record::default();
-    /// assert_eq!(record.alignment_span()?, 0);
-    /// # Ok::<_, io::Error>(())
+    /// assert_eq!(record.alignment_span(), 0);
     /// ```
-    fn alignment_span(&self) -> io::Result<u32> {
-        Ok(self.cigar().reference_len() as u32)
+    fn alignment_span(&self) -> usize {
+        self.cigar().reference_len()
     }
 
     fn mapping_quality(&self) -> Option<MappingQuality> {