+++ noodles-sam/CHANGELOG.md
@@ -51,6 +51,8 @@
 
   * sam: Rename `RecordExt` to `AlignmentRecord`.
 
+  * sam/alignment_record: Change alignment span to a `usize`.
+
   * sam/record: Move `flags`, `mapping_quality`, `read_name`,
     `template_length`, and `quality_scores` to the implementation of
     `AlignmentRecord`.